pyproject.toml
[project]
name = "athena-core"
version = "0.1.0"
description = "Autonomous Multi-Agent Framework for Program Management"
authors = [
    { name = "Team Athena", email = "team@projectathena.dev" }
]
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "pydantic>=2.6.4",
    "requests>=2.31.0",
    "python-dotenv>=1.0.1",
    "neo4j>=5.18.0",
    "py2neo>=2021.2.4",
    "pinecone>=5.0.0",
    "langgraph>=0.0.31",
    "langchain-core>=0.1.37",
    "langchain-community>=0.0.30",
    "langchain-google-genai>=1.0.1",
    "langchain-groq>=0.1.0",
    "groq>=0.4.2",
    "apscheduler>=3.10.4",
    "pytest>=8.1.1",
    "httpx>=0.27.0",
    "sqlalchemy>=2.0.29",
    "psycopg2-binary>=2.9.9"
]
requires-python = ">=3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
